DOSPERT_raw

A data.frame containing responses to the risk subscale of the Domain Specific Risk Taking Scale (DOSPERT; Weber, Blais, & Betz, 2002) based on the publicly available dataset (at https://osf.io/pjt57/) by Frey, Duncan, and Weber (2020). The items measure risk-taking propensity on five different domains: social, recreational, financial, health/ safety, and ethical.

Usage

DOSPERT_raw

Format

A data.frame with 3,123 rows (participants) and 30 columns, named by a domain prefix and item number, with six items in each of five risk domains:

ethR_1 to ethR_6

(numeric) - Ethical-domain risk-taking items.

finR_1 to finR_6

(numeric) - Financial-domain risk-taking items.

heaR_1 to heaR_6

(numeric) - Health/safety-domain risk-taking items.

recR_1 to recR_6

(numeric) - Recreational-domain risk-taking items.

socR_1 to socR_6

(numeric) - Social-domain risk-taking items.

Source

Blais, A.-R., & Weber, E. U. (2006). A domain-specific risk-taking (DOSPERT) scale for adult populations. Judgment and Decision Making, 1(1), 33–47. doi: 10.1017/S1930297500000334

Frey, R., Duncan, S. M., & Weber, E. U. (2020). Towards a typology of risk preference: Four risk profiles describe two thirds of individuals in a large sample of the U.S. population. PsyArXiv Preprint. doi:10.31234/osf.io/yjwr9