Extract a list object by its name

Extract a list object by its name

Usage

.extract_list_object(alist, object)

Arguments

alist

A list of sub-lists, typically a list of \(m\) objects of class "EFA", where \(m\) is the number of imputations passed to efa_mi().

object

String of length 1. The name of the object to extract e.g. "h2" or "vars_accounted".

Value

A list of length \(m\), with each element containing the extracted object for the \(k\)th element (\(k = 1,..., m\)).

Author

Andreas Soteriades

Consider a list of named sub-lists. This function extracts, for each sub-list, the sub-list element that is specified by the user. This function is useful for extracting results from efa_fit() for each permutation run in efa_mi().