Estimate factor scores for an EFA model

This is a wrapper function for psych::factor.scores() to be used directly with an output from EFA() or by manually specifying the factor loadings and intercorrelations. Calculates factor scores according to the specified methods if raw data are provided, and only factor weights if a correlation matrix is provided.

FACTOR_SCORES(
  x,
  f,
  Phi = NULL,
  rho = NULL,
  method = c("Thurstone", "tenBerge", "Anderson", "Bartlett", "Harman", "components")
)

Arguments

x

data.frame or matrix. Dataframe or matrix of raw data (needed to get factor scores) or matrix with correlations.

f

object of class EFA() or matrix.

Phi

matrix. A matrix of factor intercorrelations. Only needs to be specified if a factor loadings matrix is entered directly into f. Default is NULL, in which case all intercorrelations are assumed to be zero.

rho

matrix. Used when x is a matrix of raw data and the user wishes to get factor scores for a correlation matrix other than Pearson's (e.g. polychoric). Defaults to NULL, in which case psych::factor.scores() uses cor(x, use = "pairwise").

method

character. The method used to calculate factor scores. One of "Thurstone" (regression-based; default), "tenBerge", "Anderson", "Bartlett", "Harman", or "components". See psych::factor.scores() for details.

Value

A list of class FACTOR_SCORES containing the following:

scores

The factor scores (only if raw data are provided.)

weights

The factor weights.

r.scores

The correlations of the factor score estimates.

missing

A vector of the number of missing observations per subject (only if raw data are provided.

R2

Multiple R2 of the scores with the factors.

settings

A list of the settings used.

Examples

# Example with raw data with method "Bartlett"
EFA_raw <- EFA(DOSPERT_raw, n_factors = 10, type = "EFAtools", method = "PAF",
               rotation = "oblimin", randomStarts = 0)
#> ℹ `x` is not a correlation matrix; computing correlations from the raw data.
fac_scores_raw <- FACTOR_SCORES(DOSPERT_raw, f = EFA_raw, method = "Bartlett")

# Same as above, but with raw data AND a correlation matrix
cor_pearson <- cor(DOSPERT_raw)
EFA_cor_pearson <- EFA(cor_pearson, n_factors = 10, N = nrow(DOSPERT_raw),
                       type = "EFAtools", method = "PAF",
                       rotation = "oblimin", randomStarts = 0)
fac_scores_cor_pearson <- FACTOR_SCORES(DOSPERT_raw, f = EFA_cor_pearson,
                                        rho = cor_pearson,
                                        method = "Bartlett")

# Scores between two alternatives above are identical
isTRUE(all.equal(fac_scores_raw$scores, fac_scores_cor_pearson$scores,
                 check.attributes = FALSE))
#> [1] TRUE

# Example with a correlation matrix only (does not return factor scores)
EFA_cor <- EFA(test_models$baseline$cormat, n_factors = 3, N = 500,
               type = "EFAtools", method = "PAF", rotation = "oblimin")
fac_scores_cor <- FACTOR_SCORES(test_models$baseline$cormat, f = EFA_cor)
#> ℹ `x` is a correlation matrix; factor scores cannot be computed. Enter raw data
#>   to get factor scores.