Intelligence subtests from the Intelligence and Development Scales–2

A matrix containing the bivariate correlations of the 14 intelligence subtests from the Intelligence and Development Scales–2 (IDS-2; Grob & Hagmann-von Arx, 2018), an intelligence and development test battery for children and adolescents aged 5 to 20 years, for the standardization and validation sample (N = 1,991). Details can be found in Grieder & Grob (2019).

IDS2_R

Format

A 14 x 14 matrix of bivariate correlations

GS

(numeric) - Geometric shapes.

PL

(numeric) - Plates.

TC

(numeric) - Two characteristics.

CB

(numeric) - Crossing out boxes.

NL

(numeric) - Numbers / letters.

NLM

(numeric) - Numbers / letter mixed.

GF

(numeric) - Geometric figures.

RGF

(numeric) - Rotated geometric figures.

CM

(numeric) - Completing matrices.

EP

(numeric) - Excluding pictures.

CA

(numeric) - Categories.

OP

(numeric) - Opposites.

RS

(numeric) - Retelling a story.

DP

(numeric) - Describing pictures.

Source

Grieder, S., & Grob, A. (2019). Exploratory factor analyses of the intelligence and development scales–2: Implications for theory and practice. Assessment. Advance online publication. doi:10.1177/10731911198450

Grob, A., & Hagmann-von Arx, P. (2018). Intelligence and Development Scales–2 (IDS-2). Intelligenz- und Entwicklungsskalen für Kinder und Jugendliche. [Intelligence and Development Scales for Children and Adolescents.]. Bern, Switzerland: Hogrefe.