Print OMEGA object

Print OMEGA object

# S3 method for class 'OMEGA'
print(x, digits = 3, ...)

Arguments

x

output of class OMEGA (output from the OMEGA function)

digits

numeric. Passed to round. Number of digits to round to (default is 3).

...

additional arguments passed to print

Examples

efa_mod <- EFA(test_models$baseline$cormat, N = 500, n_factors = 3,
               type = "EFAtools", method = "PAF", rotation = "promax")
sl_mod <- SL(efa_mod, type = "EFAtools", method = "PAF")

OMEGA(sl_mod, type = "EFAtools",
factor_corres = sl_mod$sl[, c("F1", "F2", "F3")] >= .2)
#> Omega total, omega hierarchical, omega subscale, H index, explained common variance (ECV), and percent of uncontaminated correlations (PUC) for the general factor (top row) and omegas and H index for the group factors:
#> 
#>      tot  hier   sub     H   ECV   PUC
#> g  0.883 0.740 0.125 0.842 0.652 0.706
#> F1 0.769 0.500 0.269 0.468            
#> F2 0.763 0.494 0.270 0.477            
#> F3 0.744 0.519 0.225 0.419